Remove glib-sharp-2.0 dependency from pc file
This is not really nice, but the reason is the following:

When compiling for the OS X framework, --prefix is set to the framework location. However, when building cdnrawc, the framework is not yet installed and we try to use the .pc file for codyn-sharp to find codyn-sharp. This will fail. Instead, in cdnrawc we override pkg-config prefix variable to point to the temporary build location of codyn-sharp. This works, but pkg-config will also list any dependencies with the same overridden prefix. This will then fail for glib.